refactor `separator` into a (transformer etc.) -> transformer fn
A `make-separator` function is introduced.  Logic is expressed more in terms of
transformers ([text state] -> [text state] functions), which are a unifying
theme in this project.

This is to make it easier to modify the contents inside the delimited chunks in
arbitrary ways, as well as to modify the state.

The immediate purpose of this refactoring is to make it cleaner to add
a freezing separator, which inhibits further Markdown processing inside its
extents.

A function that walks and quacks like the old `separator` is kept around just
in case user extensions depend on it (which is tempting).