implement inhibiting separator
Unlike in the POC, here the inhibiting separator takes precedence over the code
blocks.  Although this doesn't make sense for my particular application, it's
more uniform and easier to explain/understand this way.  The inhibiting
separator can really inhibit *any* markdown parsing within a single line.

This also makes it easy to 'escape' the separator without escape markup showing
up in code blocks.  To escape a separator, just enter it twice without anything
in between.  For example, if you choose $ as a separator, and you want to
display a literal $, enter $$.  That will show up as $ anywhere, even in code
blocks, because a literal $ is output and not any escape thereof.  This
exploits the fact that `separator` will output a pair of separators verbatim if
there's not anything between them.  Thus, we can run the 'escaping' transformer
after the inhibiting transformer itself.