Inhibiting delimiters (#104)
* refactor `separator` into a (transformer etc.) -> transformer fn

A `make-separator` function is introduced.  Logic is expressed more in terms of
transformers ([text state] -> [text state] functions), which are a unifying
theme in this project.

This is to make it easier to modify the contents inside the delimited chunks in
arbitrary ways, as well as to modify the state.

The immediate purpose of this refactoring is to make it cleaner to add
a freezing separator, which inhibits further Markdown processing inside its
extents.

A function that walks and quacks like the old `separator` is kept around just
in case user extensions depend on it (which is tempting).

* allow either char seq or string as separators

Not essential, but I think we should accept the inhibiting separator as
a string, so I'm extending this to other separators.

I'm OK with rolling this part of the PR back if you don't like it.

* implement inhibiting separator

Unlike in the POC, here the inhibiting separator takes precedence over the code
blocks.  Although this doesn't make sense for my particular application, it's
more uniform and easier to explain/understand this way.  The inhibiting
separator can really inhibit *any* markdown parsing within a single line.

This also makes it easy to 'escape' the separator without escape markup showing
up in code blocks.  To escape a separator, just enter it twice without anything
in between.  For example, if you choose $ as a separator, and you want to
display a literal $, enter $$.  That will show up as $ anywhere, even in code
blocks, because a literal $ is output and not any escape thereof.  This
exploits the fact that `separator` will output a pair of separators verbatim if
there's not anything between them.  Thus, we can run the 'escaping' transformer
after the inhibiting transformer itself.

* add tests

* describe :inhibit-separator in README