Fix Text being bumped out of codeblock.
Related to this [issue](https://github.com/yogthos/markdown-clj/issues/120)
I know you said that you wanted to look at it as it could have adverse affects, but as far as I can tell this is simply fixing a typo. Content that was before \`\`\` should not  suddenly be after the \`\`\`.