Handle Hugo style anchors in heading text
The static site generator Hugo allows the following ways of defining
heading anchors in Markdown:[^1]

- `## Heading text {#heading-text-anchor}`
- `## Heading text {id=heading-text-anchor}`

When the Markdown already has these anchors built in, they are not
handled by `markdown-clj` currently.

This commit adds handling for this style of custom anchor definition
when making headings.

[^1]:
https://gohugo.io/content-management/cross-references/#heading-ids

Testing:
- Manual export of markdown content with custom anchors is working
- Manual export of markdown content without custom anchors is
working (anchor is generated)
- `clojure -M:test` all tests passed